Reorganise tests (#167)
This PR reorganises the `tests/` directory:

- move everything under `tests/golden/` (in preparation for
`tests/unit/`)
- group test files into a common directory

A side effect is that some tests that were previously not detected by
the runner now are. There is also strange cases of tests for which we
have an expectation file but no input file. I will try to figure that
out later.